qawolf.config.js: allow specifying attribute
