+++ test/qawolf.config.js
@@ -1,3 +1,4 @@
 module.exports = {
+  attribute: 'id,data-testid',
   rootDir: 'mytests',
 };
